+++ src/ios/OSCAMRParametersStruct.swift
@@ -89,3 +89,36 @@ extension OSCAMRVideoOptions {
         self.init(saveToPhotoAlbum: parameters.saveToGallery, returnMetadata: parameters.includeMetadata)
     }
 }
+
+struct OSCAMRPictureParameters: Decodable {
+    let quality: Int
+    let targetWidth: Int
+    let targetHeight: Int
+    let encodingType: Int
+    let sourceType: Int
+    let allowEdit: Bool
+    let correctOrientation: Bool
+    let saveToPhotoAlbum: Bool
+    let cameraDirection: Int 
+    let includeMetadata: Bool
+}
+
+extension OSCAMRPictureOptions {
+    convenience init(from parameters: OSCAMRPictureParameters) {
+        var targetSize = OSCAMRSize(width: parameters.targetWidth, height: parameters.targetHeight)
+        let encodingType = OSCAMREncodingType(rawValue: parameters.encodingType) ?? .jpeg
+        let direction = OSCAMRDirection(rawValue: parameters.cameraDirection) ?? .back
+
+        self.init(
+            quality: parameters.quality, 
+            size: targetSize, 
+            correctOrientation: parameters.correctOrientation, 
+            encodingType: encodingType, 
+            saveToPhotoAlbum: parameters.saveToPhotoAlbum, 
+            direction: direction, 
+            allowEdit: parameters.allowEdit, 
+            returnMetadata: parameters.includeMetadata
+        )
+        
+    }
+}